requirements.txt
absl-py==1.4.0
amqp==5.1.1
appdirs==1.4.4
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
chex==0.1.5
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.2.1
decorator==5.1.1
distrax==0.1.2
dm-env==1.6
dm-haiku==0.0.10.dev0
dm-tree==0.1.8
docker-pycreds==0.4.0
gast==0.5.3
gitdb==4.0.10
GitPython==3.1.30
grpcio==1.47.2
grpcio-tools==1.47.2
idna==3.4
imageio==2.24.0
jax==0.4.1
jaxlib==0.4.1
jmp==0.0.3
jraph==0.0.6.dev0
kombu==5.2.4
mmh3==3.0.0
numpy==1.24.1
opt-einsum==3.3.0
optax==0.1.4
pandas==1.5.2
pathtools==0.1.2
Pillow==9.4.0
prompt-toolkit==3.0.36
protobuf==3.20.0
psutil==5.9.4
pymilvus==2.2.1
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
redis==4.4.2
requests==2.28.2
rlax==0.1.5
scipy==1.10.0
sentry-sdk==1.13.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
tabulate==0.9.0
tensorflow-probability==0.19.0
toolz==0.12.0
typing_extensions==4.4.0
ujson==5.4.0
urllib3==1.26.14
vine==5.0.0
wandb==0.13.9
wcwidth==0.2.5
